% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucc_bcn_biohf_score_v2.R
\name{calc_bcn_biohf_v2}
\alias{calc_bcn_biohf_v2}
\title{Function to calculate the Barcelona BioHF Score, Version 2 (2018)}
\usage{
calc_bcn_biohf_v2(
  target_param = "1_year_mort",
  selected_model = "model1",
  age = NA,
  gender_f = NA,
  nyha = NA,
  lv_ef = NA,
  sodium = NA,
  egfr = NA,
  hb = NA,
  furosemide1 = 0,
  furosemide2 = 0,
  statin = NA,
  acei_arb = NA,
  betablock = NA,
  ntprobnp = NA,
  hstnt = NA,
  st2 = NA,
  hf_duration = NA,
  diabetes = NA,
  prev_admission = NA,
  mra = NA,
  icd = NA,
  crt = NA,
  arni = NA
)
}
\arguments{
\item{target_param}{Desired outcome of the function (n year mortality). Defaults to "1_year_mort".}

\item{selected_model}{Model to use for calculation. Has to be set according to available parameters. Defaults to "model1".}

\item{age}{Health factor patient age in years. Defaults to NA.}

\item{gender_f}{Health factor female gender encoded by f=1, m=0. Defaults to NA.}

\item{nyha}{Health factor NYHA score, can be I, II, III, IV. Defaults to NA.}

\item{lv_ef}{Health factor LV EF, Percentage. Defaults to NA.}

\item{sodium}{Health factor sodium concentration in mmol/L. Defaults to NA.}

\item{egfr}{Health factor eGFR in ml/min/m2. Defaults to NA.}

\item{hb}{Health factor hemoglobin in g/dl. Defaults to NA.}

\item{furosemide1}{Health factor medication Furosemide equivalent 0-40 mg/day encoded yes=1, no=0. Defaults to 0.}

\item{furosemide2}{Health factor medication Furosemide equivalent >40 mg/day encoded yes=1, no=0. Defaults to 0.}

\item{statin}{Health factor medication statin encoded yes=1, no=0. Defaults to NA.}

\item{acei_arb}{Health factor medication ACEi/ARB encoded yes=1, no=0. Defaults to NA.}

\item{betablock}{Health factor medication beta blockers encoded yes=1, no=0. Defaults to NA.}

\item{ntprobnp}{Health factor nt-proBNP. Defaults to NA.}

\item{hstnt}{Health factor high sensitive Troponin T. Defaults to NA.}

\item{st2}{Health factor ST2. Defaults to NA.}

\item{hf_duration}{Duration in months since when HF was first diagnosed. Only used for V2 of the score. Defaults to NA.}

\item{diabetes}{Health factor diabetes. Only used for V2 of the score. Defaults to NA.}

\item{prev_admission}{Did the patient have a HF hospitalization in the previous year? Only used for V2 of the score. Defaults to NA.}

\item{mra}{Health factor medication MRA treatment encodes yes=1, no=0. Only used for V2 of the score. Defaults to NA.}

\item{icd}{Health factor implantable cardioverter-defibrillator encodes yes=1, no=0. Only used for V2 of the score. Defaults to NA.}

\item{crt}{Health factor cardiac resynchronization therapy encodes yes=1, no=0. Only used for V2 of the score. Defaults to NA.}

\item{arni}{Health factor medication ARNi encodes yes=1, no=0. Only used for V2 of the score. Defaults to NA.}
}
\value{
The mortality probability (0-1) for the selected time window, given the selected model.
}
\description{
Function to calculate the mortality risk of patients according to
the Barcelona Bio-Heart Failure Risk Calculator (BCN Bio-HF Calculator)
which was published in two different versions.
}
\details{
This function calculates the score according to the version 2,
published in 2018.
This function requires an extended set of risk factors in comparison to
version 1, the calculations work analogues, returning a score between 0 and 1
indicating the mortality risk in the predefined timeframe.
8 different models can be used, depending on which of the risk factors
can be provided.
The algorithm was derived from the following documents of the original
publication:

Links as of 30.03.2021:

Full article:

https://onlinelibrary.wiley.com/doi/full/10.1002/ejhf.949

The web tool is available at

http://ww2.bcnbiohfcalculator.org/web/calculations

The first version of the scoring function published in 2014 can be run by using \code{calc_bcn_biohf_v1()}
}
\examples{
\dontrun{

# Load the settings data for the BCN BioHF score calculation.
# In this case these are tables with the necessary beta values and the min, max and median of the training population to replace missing values and values out of range.
data(bcn_hf_score_settings)

# The examples require dplyr package
library(dplyr)
#############################################################################

#' # Examples BCN Bio HF V2
#' # Here we use 7 additional parameters
#' # try out the function on a single patient without using it on a dataframe

calc_bcn_biohf_v2(target_param = "1_year_mort",
                          selected_model="model1",
                          age=68,
                          gender_f=0,
                          nyha="III",
                          lv_ef=30,
                          sodium=130,
                          egfr=45,
                          hb=12,
                          furosemide2=1,
                          statin=1,
                          acei_arb=1,
                          betablock=1,
                          hf_duration = 10,
                          diabetes = 0,
                          prev_admission = 1,
                          mra = 0,
                          icd = 1,
                          crt = 0,
                          arni = 0
)



# create and test a dataframe, with 4 patients:
# 1. Perfect patient, all values that are necessary
# 2. Patient missing a value, that can be imputed with median
# 3. Patient with the value that has been imputed median, expect same result
# 4. Patient with too high value, should be imputed with 99th percentile
# 5. Patient with the value that has been imputed 99th, expect same result

d_test<-data.frame( age=rep(68, 5),
                    gender_f=rep(0, 5),
                    nyha=rep("III", 5),
                    lv_ef=rep(30, 5),
                    sodium=rep(130, 5),
                    egfr=c(45,  NA, 51.2, 300, 110), #This one should be imputed
                    hb=rep(12, 5),
                    furosemide2=rep(1, 5),
                    statin=rep(1, 5),
                    acei_arb=rep(1, 5),
                    betablock=rep(1, 5),
                    hf_duration = rep(10, 5),
                    diabetes = rep(0, 5),
                    prev_admission = rep(1, 5),
                    mra = rep(0, 5),
                    icd = rep(1, 5),
                    crt = rep(0, 5),
                    arni = rep(0, 5)
)

# The function can be run using dplyr mutate and piping (\%>\%) instead of
# using it on a single patient
d_test \%>\% mutate(bcm_score=calc_bcn_biohf_v2(target_param = "3_year_mort",
                                              age = age,
                                              gender_f = gender_f,
                                              nyha = nyha,
                                              lv_ef = lv_ef,
                                              sodium = sodium,
                                              egfr = egfr,
                                              hb = hb,
                                              furosemide2 = furosemide2,
                                              statin = statin,
                                              acei_arb = acei_arb,
                                              betablock = betablock,
                                              hf_duration = hf_duration,
                                              diabetes = diabetes,
                                              prev_admission = prev_admission,
                                              mra = mra,
                                              icd = icd,
                                              crt = crt,
                                              arni = arni
)
)


# Create another dataframe, a mandatory value is missing
# 4. Patient with a value that cannot be imputed, should give a warning
d_test2<-data.frame( # age=68,
  gender_f=0,
  nyha="III",
  lv_ef=30,
  sodium=130,
  egfr=45,
  hb=12,
  furosemide2=1,
  statin=1,
  acei_arb=1,
  betablock=1,
  hf_duration = 10,
  diabetes = 0,
  prev_admission = 1,
  mra = 0,
  icd = 1,
  crt = 0,
  arni = 0
)


d_test \%>\% mutate(bcm_score=calc_bcn_biohf_v2(target_param = "3_year_mort",
                                              age = age,
                                              gender_f = gender_f,
                                              nyha = nyha,
                                              lv_ef = lv_ef,
                                              sodium = sodium,
                                              egfr = egfr,
                                              hb = hb,
                                              furosemide2 = furosemide2,
                                              statin = statin,
                                              acei_arb = acei_arb,
                                              betablock = betablock,
                                              hf_duration = hf_duration,
                                              diabetes = diabetes,
                                              prev_admission = prev_admission,
                                              mra = mra,
                                              icd = icd,
                                              crt = crt,
                                              arni = arni
)
)

d_test2 \%>\% mutate(bcm_score=calc_bcn_biohf_v2(target_param = "3_year_mort",
                                               #age = age,
                                               gender_f = gender_f,
                                               nyha = nyha,
                                               lv_ef = lv_ef,
                                               sodium = sodium,
                                               egfr = egfr,
                                               hb = hb,
                                               furosemide2 = furosemide2,
                                               statin = statin,
                                               acei_arb = acei_arb,
                                               betablock = betablock,
                                               hf_duration = hf_duration,
                                               diabetes = diabetes,
                                               prev_admission = prev_admission,
                                               mra = mra,
                                               #icd = icd,
                                               crt = crt,
                                               arni = arni
)
)

}
}
\references{
Lupón, J., Simpson, J., McMurray, J. J. V., de Antonio, M., Vila, J., Subirana, I., Barallat, J., Moliner, P., Domingo, M., Zamora, E., & Bayes-Genis, A. (2018). Barcelona Bio-HF Calculator Version 2.0: incorporation of angiotensin II receptor blocker neprilysin inhibitor (ARNI) and risk for heart failure hospitalization. European Journal of Heart Failure, 20(5), 938–940.
}
