% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucc_maggic_score.R
\name{calc_maggic_score}
\alias{calc_maggic_score}
\title{Function to calculate the MAGGIC Score (2013)}
\usage{
calc_maggic_score(
  target_param = "1_year_mort",
  age,
  lv_ef,
  sys_bp,
  bmi,
  creatinine,
  nyha,
  gender_m,
  smoking,
  diabetes,
  copd,
  hf_gt_18_months,
  betablock,
  acei_arb
)
}
\arguments{
\item{target_param}{Desired outcome of the function (n year mortality or calculated points). Defaults to "1_year_mort". Possible settings are "1_year_mort", "3_year_mort", "points" (returning the points before mapping to a mortality score), "detail_point_info" returning the string of all single scoring calculations (age, lv_ef, ...) concatenated. Order is "age|lv_ef|gender|sys_bp|bmi|creatinine|nyha|smoking|diabetes|copd|hf_gt_18_months|betablock|acei_arb".}

\item{age}{Health factor patient age in years.}

\item{lv_ef}{Health factor LV EF, Percentage.}

\item{sys_bp}{Health factor systolic blood pressure.}

\item{bmi}{Health factor body mass index.}

\item{creatinine}{Health factor creatinine in µmol/L.}

\item{nyha}{Health factor NYHA score, can be I, II, III, IV.}

\item{gender_m}{Health factor male gender encoded by m=1, f=0.}

\item{smoking}{Health factor: Is the patient a smoker? Encoded yes=1, no=0.}

\item{diabetes}{Health factor diabetes. Encoded yes=1, no=0.}

\item{copd}{Health factor COPD. Encoded yes=1, no=0.}

\item{hf_gt_18_months}{Duration in months since HF was first diagnosed is 18 months or more. Encoded yes=1, no=0.}

\item{betablock}{Health factor medication beta blockers encoded yes=1, no=0. Defaults to NA.}

\item{acei_arb}{Health factor medication ACEi/ARB encoded yes=1, no=0.}
}
\value{
The mortality probability (0-1) for the selected time window, given the selected model. (Or the the raw point information of the maggic score, depending on the target_param setting.)
}
\description{
These functions calculate the mortality risk of patients according to
the MAGGIC score, which originally uses a lookup table to assign a scoring value
to each health parameter.
This way, a final score is calculated as a sum, which can be translated to the
patient's mortality risk via a second lookup table.
The R version of the scoring algorithm was developed using the following sources:
}
\details{
Predicting survival in heart failure: a risk score based on
39 372 patients from 30 studies
Pocock et al. 2013
Eur Heart J. doi: 10.1093/eurheartj/ehs337

Figure 2: (HF risk parameters to MAGGIC points mapping)
Table 4: (MAGGIC points to 1- / 3- year mortality risk)

Links as of 21.02.2021:

Full text:
https://academic.oup.com/eurheartj/article-lookup/doi/10.1093/eurheartj/ehs337

The R function should also reconcile with the score calculator at
https://www.mdcalc.com/maggic-risk-calculator-heart-failure
}
\examples{
\dontrun{

# Load the settings data for the maggic score calculation.
# In this case it is the lookup table translating from MAGGIC score to mortality risk.
data(maggic_score_settings)

# The examples require dplyr package
library(dplyr)
#############################################################################
#Test function alone
calc_maggic_score(target_param = "1_year_mort",
                 age=68,
                 lv_ef = 30,
                 gender_m=0,
                 sys_bp=119,
                 bmi=19,
                 creatinine=135,
                 nyha="II",
                 smoking=1,
                 diabetes=0,
                 copd=0,
                 hf_gt_18_months=0,
                 betablock=1,
                 acei_arb=0)

#create and test a dataframe

d_test<-data.frame(age=68,
                  lvef = 30,
                  gender_m=0,
                  sys_bp=119,
                  bmi=19,
                  creatinine=135,
                  nyha="II",
                  smoker=1,
                  diabetes=0,
                  copd=0,
                  hf_gt_18_months=0,
                  betablock=1,
                  acei_arb=0
 )


d_test \%>\% mutate(maggic_score=calc_maggic_score(target_param = "1_year_mort",
                                                age = age,
                                                lv_ef = lvef,
                                                sys_bp = sys_bp,
                                                bmi = bmi,
                                                creatinine = creatinine,
                                                nyha = nyha,
                                                gender_m = gender_m,
                                                smoking = smoker,
                                                diabetes = diabetes,
                                                copd = copd,
                                                hf_gt_18_months = hf_gt_18_months,
                                                betablock = betablock,
                                                acei_arb = acei_arb)
                                                )

}
}
\references{
Pocock, S. J., Ariti, C. A., McMurray, J. J. V., Maggioni, A., Køber, L., Squire, I. B., Swedberg, K., Dobson, J., Poppe, K. K., Whalley, G. A., & Doughty, R. N. (2013). Predicting survival in heart failure: A risk score based on 39 372 patients from 30 studies. European Heart Journal, 34(19), 1404–1413.
}
